Initial packet parsing of RadioTap packets
Packets are switched between handler functions for:
0 - management
1 - control
2 - data
packets and some clean up in earlier code as well as changing many of the char*'s to std::string because why not use everything given to me. 

I don't look forward to parsing valid vs nonvalid frequencies and channels. Bleh

Also added better tests and logic to ignore errors as long as the interface is set correctly.